Store tp nat targets as ints, not shorts
  * In java, shorts are signed so we werent getting the full 0 - 65535 range.
  * JSON doesn't care about short vs int, everything is just an integer.

Fixes MN-2540

+++ midonet-api/src/main/java/org/midonet/api/filter/ForwardNatRule.java
@@ -76,8 +76,7 @@ public abstract class ForwardNatRule extends NatRule {
                     new org.midonet.midolman.rules.NatTarget(
                             IPv4Addr.stringToInt(natTarget.addressFrom),
                             IPv4Addr.stringToInt(natTarget.addressTo),
-                            (short) natTarget.portFrom,
-                            (short) natTarget.portTo);
+                            natTarget.portFrom, natTarget.portTo);
             targets.add(t);
         }
         return targets;
